refactor: rename cuda torch ops library files.
Update the Python executor design document to match the current CUDA, TP, JSON, and decode graph implementation.
